allow policy config to compute the policy value for account based policies (#263410)
