+++ src/vs/platform/configuration/common/configurations.ts
@@ -137,12 +137,11 @@ export class PolicyConfiguration extends Disposable implements IPolicyConfigurat
 					this.logService.warn(`Policy ${config.policy.name} has unsupported type ${config.type}`);
 					continue;
 				}
-				const { defaultValue, tags } = config.policy;
+				const { value } = config.policy;
 				keys.push(key);
 				policyDefinitions[config.policy.name] = {
 					type: config.type === 'number' ? 'number' : config.type === 'boolean' ? 'boolean' : 'string',
-					tags,
-					defaultValue,
+					value,
 				};
 			}
 		}
