+++ src/vs/platform/policy/common/policy.ts
@@ -4,14 +4,18 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IStringDictionary } from '../../../base/common/collections.js';
+import { IDefaultAccount } from '../../../base/common/defaultAccount.js';
 import { Emitter, Event } from '../../../base/common/event.js';
 import { Iterable } from '../../../base/common/iterator.js';
 import { Disposable } from '../../../base/common/lifecycle.js';
-import { PolicyName, PolicyTag } from '../../../base/common/policy.js';
+import { PolicyName } from '../../../base/common/policy.js';
 import { createDecorator } from '../../instantiation/common/instantiation.js';
 
 export type PolicyValue = string | number | boolean;
-export type PolicyDefinition = { type: 'string' | 'number' | 'boolean'; defaultValue?: string | number | boolean; tags?: PolicyTag[] };
+export type PolicyDefinition = {
+	type: 'string' | 'number' | 'boolean';
+	value?: (account: IDefaultAccount) => string | number | boolean | undefined;
+};
 
 export const IPolicyService = createDecorator<IPolicyService>('policy');
