+++ src/vs/workbench/services/policies/test/common/accountPolicyService.test.ts
@@ -4,14 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import assert from 'assert';
-import { PolicyTag } from '../../../../../base/common/policy.js';
 import { NullLogService } from '../../../../../platform/log/common/log.js';
-import { DefaultAccountService, IDefaultAccount, IDefaultAccountService } from '../../../accounts/common/defaultAccount.js';
+import { DefaultAccountService, IDefaultAccountService } from '../../../accounts/common/defaultAccount.js';
 import { AccountPolicyService } from '../../common/accountPolicyService.js';
 import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../../base/test/common/utils.js';
 import { Registry } from '../../../../../platform/registry/common/platform.js';
 import { Extensions, IConfigurationNode, IConfigurationRegistry } from '../../../../../platform/configuration/common/configurationRegistry.js';
 import { DefaultConfiguration, PolicyConfiguration } from '../../../../../platform/configuration/common/configurations.js';
+import { IDefaultAccount } from '../../../../../base/common/defaultAccount.js';
 
 const BASE_DEFAULT_ACCOUNT: IDefaultAccount = {
 	enterprise: false,
@@ -47,8 +47,7 @@ suite('AccountPolicyService', () => {
 				policy: {
 					name: 'PolicySettingB',
 					minimumVersion: '1.0.0',
-					defaultValue: "policyValueB",
-					tags: [PolicyTag.Account, PolicyTag.Preview]
+					value: account => account.chat_preview_features_enabled === false ? 'policyValueB' : undefined,
 				}
 			},
 			'setting.C': {
@@ -57,8 +56,7 @@ suite('AccountPolicyService', () => {
 				policy: {
 					name: 'PolicySettingC',
 					minimumVersion: '1.0.0',
-					defaultValue: JSON.stringify(['policyValueC1', 'policyValueC2']),
-					tags: [PolicyTag.Account, PolicyTag.Preview]
+					value: account => account.chat_preview_features_enabled === false ? JSON.stringify(['policyValueC1', 'policyValueC2']) : undefined,
 				}
 			},
 			'setting.D': {
@@ -67,8 +65,7 @@ suite('AccountPolicyService', () => {
 				policy: {
 					name: 'PolicySettingD',
 					minimumVersion: '1.0.0',
-					defaultValue: false,
-					tags: [PolicyTag.Account, PolicyTag.Preview]
+					value: account => account.chat_preview_features_enabled === false ? false : undefined,
 				}
 			},
 			'setting.E': {
