+++ src/vs/workbench/services/policies/test/common/multiplexPolicyService.test.ts
@@ -4,9 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import assert from 'assert';
-import { PolicyTag } from '../../../../../base/common/policy.js';
 import { NullLogService } from '../../../../../platform/log/common/log.js';
-import { DefaultAccountService, IDefaultAccount, IDefaultAccountService } from '../../../accounts/common/defaultAccount.js';
+import { DefaultAccountService, IDefaultAccountService } from '../../../accounts/common/defaultAccount.js';
 import { AccountPolicyService } from '../../common/accountPolicyService.js';
 import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../../base/test/common/utils.js';
 import { Registry } from '../../../../../platform/registry/common/platform.js';
@@ -19,6 +18,7 @@ import { IFileService } from '../../../../../platform/files/common/files.js';
 import { InMemoryFileSystemProvider } from '../../../../../platform/files/common/inMemoryFilesystemProvider.js';
 import { FileService } from '../../../../../platform/files/common/fileService.js';
 import { VSBuffer } from '../../../../../base/common/buffer.js';
+import { IDefaultAccount } from '../../../../../base/common/defaultAccount.js';
 
 const BASE_DEFAULT_ACCOUNT: IDefaultAccount = {
 	enterprise: false,
@@ -56,8 +56,7 @@ suite('MultiplexPolicyService', () => {
 				policy: {
 					name: 'PolicySettingB',
 					minimumVersion: '1.0.0',
-					defaultValue: "policyValueB",
-					tags: [PolicyTag.Account, PolicyTag.Preview]
+					value: account => account.chat_preview_features_enabled === false ? 'policyValueB' : undefined,
 				}
 			},
 			'setting.C': {
@@ -66,8 +65,7 @@ suite('MultiplexPolicyService', () => {
 				policy: {
 					name: 'PolicySettingC',
 					minimumVersion: '1.0.0',
-					defaultValue: JSON.stringify(['policyValueC1', 'policyValueC2']),
-					tags: [PolicyTag.Account, PolicyTag.Preview]
+					value: account => account.chat_preview_features_enabled === false ? JSON.stringify(['policyValueC1', 'policyValueC2']) : undefined,
 				}
 			},
 			'setting.D': {
@@ -76,8 +74,7 @@ suite('MultiplexPolicyService', () => {
 				policy: {
 					name: 'PolicySettingD',
 					minimumVersion: '1.0.0',
-					defaultValue: false,
-					tags: [PolicyTag.Account, PolicyTag.Preview]
+					value: account => account.chat_preview_features_enabled === false ? false : undefined,
 				}
 			},
 			'setting.E': {
